pyproject.toml
[tool.poetry]
name = "fastapiregistration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = "^0.19.3"
fastapi = { extras = ["all"], version = "^0.92.0" }
uvicorn = { extras = ["standard"], version = "^0.20.0" }
aerich = "^0.7.1"
asyncpg = "^0.27.0"
black = "^23.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.aerich]
tortoise_orm = "app.config.db_config.DB_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
